RWLocks, cap group concurrency
Made some tweaks to use RWLocks between pick and
sort functions.

Updated main lock block to initially use TryLock.
This is mainly to protected against a hypothetical
race condition where a second connection could get
the same plot before the first locked it. Its
hypothetical and would self resolve once the first
transfer was done, but prefer to check and return
for the client to retry vs block it.

Ensure plotGroup concurrency does not exceed the
number of drives.